Intent: strengthen cache API coverage Operators involved: [Coherence, Recursivity] Affected invariants: [#7] Key changes: - extend dnfr cache tests to cover cached_nodes_and_A and cached_node_list via public APIs - streamline cache helper tests to avoid private symbol usage Expected risks/dissonances: minimal; public API coverage now shoulders cache integrity Metrics: functional coverage maintained via pytest suite Equivalence map: none
